Move EventTarget into utilities
