+++ lib/utilities/attribute.js
@@ -3,7 +3,7 @@
 
 var ObjectUtils = require('./objects');
 var Lang = require('./lang');
-var EventTarget = require('./../eventTarget');
+var EventTarget = require('./eventTarget');
 var augment = require('./augment');
 
 var OPTS = {
@@ -28,6 +28,13 @@ _writtenOnce - whether it's been set at least once
 
 var VALID_OPTIONS = ObjectUtils.values(OPTS);
 
+/**
+ * Perform sanity checks on an attr config. 
+ * Reject unknown parameters.
+ * @method validateConfig
+ * @param  {Object} cfg
+ * @return {Boolean} is valid
+ */
 var validateConfig = function(cfg) {
   for (var key in cfg) {
     if (cfg.hasOwnProperty(key) &&
@@ -45,6 +52,12 @@ var validateConfig = function(cfg) {
   return true;
 };
 
+/**
+ * Get an instance method on this by name, logging an error if it does not exist.
+ * @method stringToInstanceMethod
+ * @param  {String} functionName
+ * @return {Function}
+ */
 var stringToInstanceMethod = function(functionName) {
   if (!Lang.isFunction(this[functionName])) {
     console.error('Instance method function ' + functionName + ' does not exist.');
@@ -79,6 +92,13 @@ var Attribute = function() {
 
 Attribute.prototype = {
 
+  /**
+   * Configure several attributes at once. 
+   * Accepts an object mapping attribute names to config objects.
+   * See #addAttr for detail on the config object.
+   * @method addAttrs
+   * @param  {[type]} configs
+   */
   addAttrs: function(configs) {
     ObjectUtils.each(configs, function(config, name) {
       this.addAttr(name, config);
@@ -140,6 +160,12 @@ Attribute.prototype = {
     this._attrStates[name] = {};
   },
 
+  /**
+   * Throw an error if an attribute with name does not exist.
+   * @method  _validateHasAttr
+   * @param   {String} name
+   * @private
+   */
   _validateHasAttr: function(name) {
     var config = this._attrConfigs[name];
