+++ lib/utilities/eventTarget.js
@@ -1,5 +1,5 @@
-var ObjectUtils = require('./utilities/objects');
-var Lang = require('./utilities/lang');
+var ObjectUtils = require('./objects');
+var Lang = require('./lang');
 /**
  * A simple implementation of the observer pattern.
  * Mix me into your class with augment.
